During the creeation of the vivado project the block design is now created via a script.
Before it was imported via the .bd file. However, this somehow caused problems because IP cores were miss configured and this never threw an error and the systhesis and implementation finished with out problems. But, the design did not behave as it did previously when it was build in the original project.